+++ src/versioningit/errors.py
@@ -1,3 +1,6 @@
+from pathlib import Path
+
+
 class Error(Exception):
     """Base class of all ``versioningit``-specific errors"""
 
@@ -25,6 +28,39 @@ class NotVersioningitError(Error):
     """
 
     pass
+
+
+class NoConfigFileError(NotVersioningitError):
+    """
+    .. versionadded:: 3.2.0
+
+    Raised when ``versioningit`` is used on a project that does not contain a
+    :file:`versioningit.toml` or :file:`pyproject.toml` file
+    """
+
+    def __init__(self, project_dir: Path) -> None:
+        #: The path to the project directory
+        self.project_dir: Path = project_dir
+
+    def __str__(self) -> str:
+        return f"No pyproject.toml or versioningit.toml file in {self.project_dir}"
+
+
+class NoConfigSectionError(NotVersioningitError):
+    """
+    .. versionadded:: 3.2.0
+
+    Raised when ``versioningit`` is used on a project whose
+    :file:`versioningit.toml` or :file:`pyproject.toml` file does not contain a
+    ``versioningit`` configuration table
+    """
+
+    def __init__(self, config_path: Path) -> None:
+        #: The path to the configuration file
+        self.config_path: Path = config_path
+
+    def __str__(self) -> str:
+        return f"versioningit not configured in {self.config_path}"
 
 
 class NotSdistError(Error):